Remove RelationSpec case insensitivity
This mostly feels like a gotcha. It is admittedly nice for querying for
specific relations but it feels quite counter intuitive that upsertion
can result in a Relation that doesn't match the case of RelationSpec if
it already existed with a different case.

The API is meant to give you a playground to create your own relations
and operate on them not make up assumptions such as casing insensitivity