Rename upsert_relation to obtain_relation
Upsert was not the correct verb here. This is NOT an update OR insert.
And I don't think it ever will be.
Upsert here would imply that if you were to add a "Due date" relation
with format text and we found an existing one we will update its format
to text, this is not a behavior I will personally ever want so I am not
interested in upsertion, I am interested in "get or create" which I will
give the name obtain from here on out