Resolve recommended_relations for ObjectType
Returning RelationIds unresolved will eventually lead to problems
because we don't know whether a relation is hidden or not without
querying it. So instead we resolve ObjectType's recommended_relations
field before handing it back to the user

This does not make us do any more requests than we were already doing,
it's simply handing the user some of the internal information we were
throwing away after validating it. It would cause us to need to do more
requests only if we implement Space::get_object_type_by_id but that's
an okay trade-off.