Use a list for SearchOutput::Layout
Objects can have many different layouts, for them layout is but a
property. For example built into the default clients you can choose the
layout of objects as one of four things:
- Basic
- Profile
- Action
- Note

But also some builtin objects have layouts that are just an object with
a different layout anyway, for instance Bookmark. For the four above
layouts they really all belong to the same Object. For something like
Bookmark a specialized type might be better but for now I think it's
fair to treat it like any other Object