Remove is_hidden fn from SearchOutput trait
Didn't need to be a method for each implementation. Since everything is
fully relying on object_search gRPC call at this point, we can use the
fact we only call it from one function to our advantage. We will get the
`isHidden` field in Space::search_object and use it early