Add slow_resolve method to ObjectType
The recommended relations aren't always available in the context when
creating a new ObjectType, unfortunately. Starting to notice a bit
of a pattern here though where every SearchOutput is able to resolve
into some other type with the help of a `Space`. Probably worth to
capture that at some point after we add some per external call caching
to `Space`